More explicit information about the use of covariates in models.
If you are using GEM, you do NOT need to include the exposure as a covariate, since GEM automatically includes it for you. If you try to include it, you will get an error.

If you are using MAGEE, you MUST include the exposure as a covariate! It does not get automatically included.